Started adding TF-IDF to tokenizer. We could try to calculate tf in the tokenizer, then pass it, another with other data to MapBuilder. MapBuilder would use this data to calculate the idf portion.
We can sum the term frequencies, for duplicate words, then multiply it by idf later.

tf * idf + tf * idf + tf * idf --> (tf + tf + tf) * idf

I'm not sure how else to calculate tf-idf at a sentence level, which the requirement for this assignment.
I spoke to our instructor about this, and they told me that we're supposed to calculate it this way.

tf = the count of some word / all words in a sentence
idf = the count of all sentences / the count of sentences with some word

At some point, we'd have words that appear in more than one sentence, thus multiple tf values for one word.
I think that we should add them, but I'm open to suggestions.