I decided to make the TF-IDF process a separate class, which is used prior to tokenization. I tried to incorporate it into the MapBuilder class and the Tokenizer, but I ran into several complications.
With the MapBuilder, we wouldn't have the sentence structure necessary to calculate tf.
With the tokenizer, we'd have to recaculate metrics when we delete words from our list.

Although we would have to make multiple passes through the data, it's easiest to remove the words prior to tokenization.
That way, the tokenizer could calculate metrics that we wouldn't have to recalculate and the MapBuilder could
use whatever tokens that it is given.

Now, the TF-IDF class also separates the test data from the training data.