Made some changes to how TFIDF, & logistic regression were calculated. Tried to streamline the activity of the preprocessor.
I moved a lot of data / statistical analysis stuff to the pr folder.

The build.sh script in the pr folder runs the project in 'debug' mode, where it prints
tfidf scores & ngrams to disk.

The gram-freq script generates frequencies of the n-grams in the preprocesor folder &
sorts the tfidf-output.